requirements.txt
flask==3.1.3
pillow==12.2.0
gunicorn==25.3.0